requirements.txt
absl-py==2.1.0
addict==2.4.0
astunparse==1.6.3
basicsr==1.4.2
blinker==1.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
facexlib==0.3.0
filterpy==1.4.5
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
future==1.0.0
gast==0.6.0
gfpgan==1.3.8
google-pasta==0.2.0
grpcio==1.67.1
gunicorn==23.0.0
h5py==3.12.1
idna==3.4
imageio==2.36.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.3.3
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
networkx==3.4.2
numba==0.60.0
numpy==1.26.3
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==10.2.0
platformdirs==4.3.6
protobuf==4.25.5
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
realesrgan==0.3.0
requests==2.28.1
rich==13.9.4
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
tb-nightly==2.19.0a20241112
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2024.9.20
tomli==2.1.0
torch==1.13.1+cpu
torchvision==0.14.1+cpu
tqdm==4.67.0
typing_extensions==4.9.0
urllib3==1.26.13
Werkzeug==3.1.3
wrapt==1.16.0
yapf==0.40.2
zipp==3.21.0
